# cedar-bridge-server-application/src/main/resources/assets/swagger-api/swagger.yaml
openapi: 3.0.1
info:
  description: |
    Better Metadata - Better Science

    The bridge server is CEDAR's connection to registries outside it. It mints DOIs through DataCite
    so a CEDAR artifact can be cited, and it resolves and searches identifiers held by seven external
    authorities: DOI, NIH grants, ORCID, CompTox, PubMed, ROR and RRID.

    Where an answer comes from outside, the status comes from outside with it: an upstream refusal is
    reported as that registry reported it, rather than flattened into a 500.
  title: CEDAR Bridge Server API
  version: 2.9.19
paths:
...
  /ext-auth/{authority}/search-by-name:
    get:
      description: "These routes take no credentials. Neither builds a request context,\
        \ so anyone who can reach this host can use them, and three of the seven authorities\
        \ behind them spend credentials the deployment holds. Recorded here because\
        \ a spec that claimed otherwise would be worse than one that says so. Search\
        \ one external authority for entries matching a name, and return them with\
        \ the paging that produced them. The status is the authority's own, so an\
        \ upstream refusal is reported as that authority reported it. An authority\
        \ that has not finished loading answers 503 with Retry-After rather than an\
        \ empty result."
      operationId: searchByName
      parameters:
      - description: "Which registry to ask. One of `doi`, `nih-grant`, `orcid`, `comp-tox`,\
          \ `pmid`, `ror`, `rrid`. A segment no authority is registered under answers\
          \ 404 naming the ones that are."
        in: path
        name: authority
        required: true
        schema:
          type: string
      - description: The name to search for.
        in: query
        name: q
        schema:
          type: string
      - description: "How many entries to return, from 1 to 500. Defaults to 100."
        in: query
        name: limit
        schema:
          type: integer
          format: int32
      - description: How many matching entries to skip. Defaults to 0.
        in: query
        name: offset
        schema:
          type: integer
          format: int32
      - description: "Zero-based page number, for clients that page by number. Cannot\
          \ be sent with `limit` or `offset`. Defaults to 0."
        in: query
        name: page
        schema:
          type: integer
          format: int32
      - description: "Entries per page, for clients that page by number. Cannot be\
          \ sent with `limit` or `offset`. Defaults to 100, and must be greater than\
          \ one and at most 500."
        in: query
        name: pageSize
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthoritySearchResults"
          description: "A page of matching entries, with the paging envelope and,\
            \ for clients that still page by number, `found`, `page` and `pageSize`"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: "The limit or offset is out of range, `page` is negative or\
            \ `pageSize` is not greater than one, or both kinds of paging were sent"
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Unauthorized
        "404":
          description: No authority is served under this path segment
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: The authority is not ready yet; Retry-After says when to try
            again
      summary: Search an external registry by name
      tags:
      - External authorities
  /ext-auth/{authority}/{id}:
    get:
      description: "These routes take no credentials. Neither builds a request context,\
        \ so anyone who can reach this host can use them, and three of the seven authorities\
        \ behind them spend credentials the deployment holds. Recorded here because\
        \ a spec that claimed otherwise would be worse than one that says so. Look\
        \ one identifier up in an external authority and return what it holds, with\
        \ `found` saying whether it resolved and `requestedId` echoing what was asked.\
        \ The status is the authority's own. This path and the search path both match\
        \ two segments; the literal `search-by-name` wins, so no authority can have\
        \ an entry by that name."
      operationId: details
      parameters:
      - description: "Which registry to ask. One of `doi`, `nih-grant`, `orcid`, `comp-tox`,\
          \ `pmid`, `ror`, `rrid`. A segment no authority is registered under answers\
          \ 404 naming the ones that are."
        in: path
        name: authority
        required: true
        schema:
          type: string
      - description: "The identifier to resolve, as that registry spells it."
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorityDetails"
          description: What the authority holds for the identifier
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: Unauthorized
        "404":
          description: "No authority is served under this path segment, or the authority\
            \ does not hold this identifier"
        "503":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CedarError"
          description: The authority is not ready yet; Retry-After says when to try
            again
      summary: Resolve an identifier against an external registry
      tags:
      - External authorities
...
components:
  schemas:
    AuthorityDetails:
      type: object
      additionalProperties: true
      description: |
        What one authority holds for one identifier. Only `found` and `requestedId` are added by
        CEDAR; the rest of the object is the authority's own, and the keys below are the ones every
        authority sets when it resolves an identifier. ORCID and ROR also pass the registry's whole
        record through under `rawResponse`, and its errors under `errors`.
      properties:
        details:
          type: string
          description: "A line of supporting detail, where the authority composes\
            \ one."
        errors:
          type: array
          description: "What the registry said went wrong, one message per entry."
          items:
            type: string
        found:
          type: boolean
          description: Whether the identifier resolved.
        id:
          type: string
          format: uri
          description: "The entry's IRI, as the authority spells it."
          nullable: true
        name:
          type: string
          description: The entry's display name. Null where the identifier did not
            resolve.
          nullable: true
        rawResponse:
          type: object
          additionalProperties: true
          description: "The registry's own record, unchanged, where the authority\
            \ passes it through."
        requestedId:
          type: string
          description: The identifier as it was asked for.
      required:
      - found
      - requestedId
    AuthoritySearchRequest:
      type: object
      additionalProperties: false
      description: "The page that was asked for, after defaults were applied."
      properties:
        limit:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
      required:
      - limit
      - offset
    AuthoritySearchResults:
      type: object
      additionalProperties: false
      description: |
        The entries one authority offers for a name, in CEDAR's paging envelope. The status is the
        authority's own, so an upstream refusal arrives with the registry's status and its reason
        under `errors`. `found`, `page` and `pageSize` are the fields the route answered with before
        it took `limit` and `offset`, kept for clients that still page by number.
      properties:
        countCapped:
          type: boolean
          description: Present and true when `totalCount` is a lower bound. There
            is then no `last` link.
        currentOffset:
          type: integer
          format: int64
          description: The offset of the first entry on this page.
        errors:
          type: array
          description: |
            What the registry said went wrong, one message per entry. Present only where the registry
            reported something; only ORCID does.
          items:
            type: string
        found:
          type: boolean
          description: Whether the registry answered and offered at least one entry.
        page:
          type: integer
          format: int32
          description: "The zero-based page returned, defaulted to 0."
        pageSize:
          type: integer
          format: int32
          description: "The page size used, defaulted to 100."
        paging:
          type: object
          additionalProperties:
            type: string
          description: "Links to the first, previous, next and last pages, keyed by\
            \ relation. A relation with no page is absent."
        request:
          $ref: "#/components/schemas/AuthoritySearchRequest"
        results:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/AuthorityTerm"
          description: "One entry per matching term, keyed by the term's IRI, in the\
            \ order the registry offered them."
        totalCount:
          type: integer
          format: int64
          description: |
            How many entries match. A lower bound when `countCapped` is true: ROR can be read only
            one upstream page deep, and NIH RePORTER narrows its answers after the fact and cannot
            know its total at all.
      required:
      - currentOffset
      - found
      - page
      - pageSize
      - paging
      - request
      - results
      - totalCount
    AuthorityTerm:
      type: object
      additionalProperties: false
      description: One entry an authority offers for a name.
      properties:
        details:
          type: string
          description: |
            A line of supporting detail the authority composes to tell one entry from another of the
            same name. Its wording is the authority's own.
        name:
          type: string
          description: "The entry's display name, chosen by the authority from the\
            \ names the registry holds."
    CedarError:
      type: object
      additionalProperties: false
      description: A CEDAR error response. Diagnostic fields are populated only when
        they are relevant to the failure.
      properties:
        entities:
          type: object
          additionalProperties:
            type: object
            description: Named domain entities associated with exception-mapped errors.
            nullable: true
          description: Named domain entities associated with exception-mapped errors.
          nullable: true
        errorId:
          type: string
          format: uuid
          description: Correlation identifier for an internal error recorded in server
            logs.
          nullable: true
        errorKey:
          type: string
          description: Stable machine-readable error identifier.
          enum:
          - templateElementNotCreated
          - templateElementNotFound
          - templateElementNotDeleted
          - templateElementNotUpdated
          - templateElementsNotListed
          - templateFieldNotCreated
          - templateFieldNotFound
          - templateFieldNotDeleted
          - templateFieldNotUpdated
          - templateFieldsNotListed
          - templateNotCreated
          - templateNotFound
          - templateNotDeleted
          - templateNotUpdated
          - templatesNotListed
          - templateInstanceNotCreated
          - templateInstanceNotFound
          - templateInstanceNotDeleted
          - templateInstanceNotUpdated
          - templateInstancesNotListed
          - noReadAccessToResource
          - noWriteAccessToResource
          - noReadAccessToFolder
          - noWriteAccessToFolder
          - noReadAccessToTemplate
          - noWriteAccessToTemplate
          - noReadAccessToTemplateElement
          - noWriteAccessToTemplateElement
          - noReadAccessToTemplateField
          - noWriteAccessToTemplateField
          - noReadAccessToTemplateInstance
          - noWriteAccessToTemplateInstance
          - noReadAccessToArtifact
          - noWriteAccessToArtifact
          - noReadAccessToCategory
          - noWriteAccessToCategory
          - folderNotFound
          - artifactNotFound
          - artifactPreconditionRequired
          - artifactHasMovedOn
          - verbatimWriteRefused
          - nodeNotFound
          - unknownResourceType
          - missingParameter
          - invalidInput
          - sourceFolderNotFound
          - sourceResourceNotFound
          - targetFolderNotFound
          - nodeNotMoved
          - groupNotFound
          - userNotFound
          - uniqueConstraintCollision
          - invalidData
          - notAuthorized
          - permissionMissing
          - tokenInvalid
          - userInfoLoadByTokenFailed
          - userInfoLoadByApiKeyFailed
          - cedarUserNotFound
          - authorizationNotFound
          - permissionNotOwned
          - authorizationTypeUnknown
          - apiKeyNotFound
          - tokenMissing
          - tokenExpired
          - parentFolderNotSpecified
          - parentFolderSpecifiedTwice
          - pathNotNormalized
          - parentFolderNotFound
          - updateInvalidFolderName
          - createInvalidFolderName
          - nodeAlreadyPresent
          - folderNotCreated
          - missingNameAndDescription
          - missingData
          - folderNotDeleted
          - artifactNotDeleted
          - resourceNotCreated
          - invalidResourceType
          - invalidArtifactType
          - readOtherProfileForbidden
          - updateOtherProfileForbidden
          - folderCanNotBeDeleted
          - folderCanNotBeChanged
          - groupAlreadyPresent
          - groupCanBeModifiedOnlyByGroupAdmin
          - groupCanBeDeletedOnlyByGroupAdmin
          - groupMembersCanBeReadOnlyByGroupAdmin
          - groupRequiresAdministrator
          - groupUsersNotUpdated
          - specialGroupCanNotBeDeleted
          - folderPermissionsCanNotBeChanged
          - unknownInstanceOutputFormat
          - folderCopyNotAllowed
          - methodNotImplemented
          - upstreamServerError
          - nothingToDo
          - parentCategoryNotFound
          - categoryAlreadyPresent
          - categoryNotFound
          - categoryCanNotBeDeleted
          - rootCategoryCanNotBeDeleted
          - unableToAttachCategory
          - unableToDetachCategory
          - noCategoriesWereAttached
          - malformedJsonRequestBody
          - malformedSearchTerm
          - publishedArtifactCanNotBeChanged
          - publishedArtifactCanNotBeDeleted
          - versioningOnlyOnLatest
          - versioningOnlyByOwner
          - nonVersionedArtifactType
          - createDraftOnlyFromPublished
          - publishOnlyDraft
          - draftNotCreated
          - contentNotValid
          - doiNotSupportedByResourceType
          - doiCanNotBeSetForEmptyAtId
          - doiCanNotBeAltered
          - doiCanNotBeSet
          - doiAlreadyExists
          - dataCiteDOIDisabled
          - pinnedVersionUnavailable
          - resourceNotFound
          - internalError
          nullable: true
        errorReasonKey:
          type: string
          description: More specific machine-readable reason for the error.
          enum:
          - nonEmptyFolder
          - nonEmptyCategory
          - userHomeFolder
          - systemFolder
          - templateReferencedInInstances
          - validationError
          - continuationExpired
          nullable: true
        errorType:
          type: string
          description: Broad category of the error.
          enum:
          - notFound
          - invalidArgument
          - authorization
          - permission
          - server
          - validationError
          nullable: true
        message:
          type: string
          description: Human-readable error message.
          nullable: true
        objects:
          type: object
          additionalProperties:
            type: object
            description: Named structured values associated with the error.
          description: Named structured values associated with the error.
        operation:
          type: object
          description: Operation that failed. Its fields depend on the operation type.
          nullable: true
        parameters:
          type: object
          additionalProperties:
            type: object
            description: Named scalar values associated with the error.
          description: Named scalar values associated with the error.
        status:
          type: string
          description: Symbolic HTTP response status.
          enum:
          - BAD_REQUEST
          - UNAUTHORIZED
          - FORBIDDEN
          - NOT_FOUND
          - METHOD_NOT_ALLOWED
          - NOT_ACCEPTABLE
          - CONFLICT
          - PRECONDITION_FAILED
          - UNSUPPORTED_MEDIA_TYPE
          - UNPROCESSABLE_ENTITY
          - PRECONDITION_REQUIRED
          - INTERNAL_SERVER_ERROR
          - NOT_IMPLEMENTED
          - BAD_GATEWAY
          - SERVICE_UNAVAILABLE
          - HTTP_VERSION_NOT_SUPPORTED
        statusCode:
          type: integer
          format: int32
          description: Numeric HTTP response status.
          maximum: 599
          minimum: 400
        suggestedAction:
          type: string
          description: Recovery action suggested to the client.
          enum:
          - none
          - requestRole
          - logout
          - logoutImmediately
          - provideAuthorizationHeader
          - refreshToken
          nullable: true
      required:
      - status
      - statusCode
...
  securitySchemes:
    api_key:
      description: "Your CEDAR API key, using the format 'apiKey <yourApiKey>'"
      in: header
      name: Authorization
      type: apiKey